Crate test_casing

Minimalistic framework for generating tests for a given set of test cases. In other words, it implements parameterized tests of reasonably low cardinality for the standard Rust test runner.

Overview

The core export of this crate is the test_casing attribute macro. It wraps a free-standing function with one or more arguments and transforms it into a collection of test cases. The arguments to the function are supplied by an iterator (more precisely, an expression implementing IntoIterator).

For convenience, there is TestCases, a lazy iterator wrapper that allows constructing test cases which cannot be constructed in compile time (e.g., ones requiring access to heap). TestCases can be instantiated using the cases! macro.

Intended use cases

Since a separate test wrapper is generated for each case, their number should be reasonably low (roughly speaking, no more than 20).

Isolating each test case makes most sense if the cases involve some heavy lifting (spinning up a runtime, logging considerable amount of information, etc.).

Test cases structure

The generated test cases are placed in a module with the same name as the target function near the function. This allows specifying the (potentially qualified) function name to restrict the test scope.

If the nightly crate feature is not enabled, names of particular test cases are not descriptive; they have the case_NN format, where NN is the 0-based case index. The values of arguments provided to the test are printed to the standard output at the test start. (The standard output is captured and thus may be not visible unless the --nocapture option is specified in the cargo test command.)

If the nightly feature is enabled, the names are more descriptive, containing Debug presentation of all args together with their names. Here’s an excerpt from the integration tests for this crate:

test number_can_be_converted_to_string::case_1 [number = 3, expected = "3"] ... ok
test number_can_be_converted_to_string::case_2 [number = 5, expected = "5"] ... ok
test numbers_are_large::case_0 [number = 2] ... ignored, testing that `#[ignore]` attr works
test numbers_are_large::case_1 [number = 3] ... ignored, testing that `#[ignore]` attr works
test string_conversion_fail::case_0 [bogus_str = "not a number"] - should panic ... ok
test string_conversion_fail::case_1 [bogus_str = "-"] - should panic ... ok
test string_conversion_fail::case_2 [bogus_str = ""] - should panic ... ok

The names are fully considered when filtering tests, meaning that it’s possible to run particular cases using a filter like cargo test 'number = 5'.

Alternatives and similar tools

The approach from this crate can be reproduced with some amount of copy-pasting by manually feeding necessary inputs to a common parametric testing function. Optionally, these tests may be collected in a module for better structuring. The main downside of this approach is the amount of copy-pasting.

Alternatively, multiple test cases may be run in a single #[test] (e.g., in a loop). This is fine for the large amount of small cases (e.g., mini-fuzzing), but may have downsides such as overflowing or overlapping logs and increased test runtimes.

The test-case crate uses a similar approach to test case structuring, but differs in how test case inputs are specified. Subjectively, the approach used by this crate is more extensible and easier to read.

Property testing / quickcheck-like frameworks provide much more exhaustive approach to parameterized testing, but they require significantly more setup effort.

Crate features

nightly

(Off by default)

Uses custom test frameworks APIs together with a generous spicing of hacks to include arguments in the names of the generated tests (see an excerpt above for an illustration). test_casing actually does not require a custom test runner, but rather hacks into the standard one; thus, the generated test cases can run alongside with ordinary / non-parameterized tests.

Requires a nightly Rust toolchain and specifying #![feature(test, custom_test_frameworks)] in the using crate.

Examples

Basic usage

test_casing macro accepts 2 args: number of cases and the iterator expression. The latter can be any valid Rust expression.

#[test_casing(5, 0..5)]
// #[test] attribute is optional and is added automatically
// provided that the test function is not `async`.
fn number_is_small(number: i32) {
    assert!(number < 10);
}

Functions returning Results are supported as well.

use std::error::Error;

#[test_casing(3, ["0", "42", "-3"])]
fn parsing_numbers(s: &str) -> Result<(), Box<dyn Error>> {
    let number: i32 = s.parse()?;
    assert!(number.abs() < 100);
    Ok(())
}

The function on which the test_casing attribute is placed can be accessed from other code (e.g., for more tests):

#[test_casing(3, ["0", "42", "-3"])]
fn parsing_numbers(s: &str) -> Result<(), Box<dyn Error>> {
    // snipped...
}

#[test]
fn parsing_number_error() {
    assert!(parsing_numbers("?").is_err());
}

Case expressions

Case expressions can be extracted to a constant for reuse or better code structuring.

const CASES: TestCases<(String, i32)> = cases! {
    [0, 42, -3].map(|i| (i.to_string(), i))
};

#[test_casing(3, CASES)]
fn parsing_numbers(s: String, expected: i32) {
    let parsed: i32 = s.parse().unwrap();
    assert_eq!(parsed, expected);
}

This example also shows that semantics of args is up to the writer; some of the args may be expected values, etc.

Cartesian product

One of possible case expressions is a Product; it can be used to generate test cases as a Cartesian product of the expressions for separate args.

#[test_casing(6, Product((0_usize..3, ["foo", "bar"])))]
fn numbers_and_strings(number: usize, s: &str) {
    assert!(s.len() <= number);
}

Reference args

It is possible to go from a generated argument to its reference by adding a #[map(ref)] attribute on the argument. The attribute may optionally specify a path to the transform function from the reference to the desired type (similar to transform specifications in the serde attr).

const CASES: TestCases<(String, i32)> = cases! {
    [0, 42, -3].map(|i| (i.to_string(), i))
};

#[test_casing(3, CASES)]
fn parsing_numbers(#[map(ref)] s: &str, expected: i32) {
    // Snipped...
}

#[test_casing(3, CASES)]
fn parsing_numbers_too(
    #[map(ref = String::as_str)] s: &str,
    expected: i32,
) {
    // Snipped...
}

ignore and should_panic attributes

ignore or should_panic attributes can be specified below the test_casing attribute. They will apply to all generated tests.

#[test_casing(3, ["not", "implemented", "yet"])]
#[ignore = "Promise this will work sometime"]
fn future_test(s: &str) {
    unimplemented!()
}

#[test_casing(3, ["not a number", "-", ""])]
#[should_panic(expected = "ParseIntError")]
fn string_conversion_fail(bogus_str: &str) {
    bogus_str.parse::<i32>().unwrap();
}

Async tests

test_casing supports all kinds of async test wrappers, such as async_std::test, tokio::test, actix::test etc. The corresponding attribute just needs to be specified below the test_casing attribute.

#[test_casing(3, ["0", "42", "-3"])]
#[async_std::test]
async fn parsing_numbers(s: &str) -> Result<(), Box<dyn Error>> {
    assert!(s.parse::<i32>()?.abs() < 100);
    Ok(())
}

Macros

cases

Creates TestCases based on the provided expression implementing IntoIterator (e.g., an array, a range or an iterator).

Structs

Product

Cartesian product of several test cases.

ProductIter

Iterator over test cases in Product.

TestCases

Container for test cases based on a lazily evaluated iterator. Should be constructed using the cases! macro.

Attribute Macros

test_casing

Flattens a parameterized test into a collection of test cases.