Attribute Macro test_casing::test_casing

#[test_casing]

Flattens a parameterized test into a collection of test cases.

Inputs

This attribute must be placed on a free-standing function with 1..8 arguments. The attribute must be invoked with 2 values:

Number of test cases, a number literal

A case iterator expression evaluating to an implementation of IntoIterator with Debuggable, 'static items. If the target function has a single argument, the iterator item type must equal to the argument type. Otherwise, the iterator must return a tuple in which each item corresponds to the argument with the same index.

A case iterator expression may reference the environment (e.g., it can be a name of a constant). It doesn’t need to be a constant expression (e.g., it may allocate in heap). It should return at least the number of items specified as the first attribute argument, and can return more items; these additional items will not be tested.

Mapping arguments

To support more idiomatic signatures for parameterized test functions, it is possible to map from the type returned by the case iterator. The only supported kind of mapping so far is taking a shared reference (i.e., T → &T). The mapping is enabled by placing the #[map(ref)] attribute on the corresponding argument. Optionally, the reference &T can be further mapped with a function / method (e.g., &String → &str with String::as_str()). This is specified as #[map(ref = path::to::method)], a la serde transforms.

Examples

See test-casing crate-level docs for the examples of usage.